requirements.txt
pillow
pandas==1.2.4
numpy
scikit-learn
opencv-python==4.5.3.56
monai==0.8.0
efficientnet-pytorch==0.7.1
torch==1.10.0
torchaudio==0.10.0
torchinfo==1.5.3
torchvision==0.11.1
